Initial commit. Semi-functional.
MTP is horrible.
Everything in here that works uses the precompiled LibMTP
Example program.
Some MTP functionality works, like querying device state.
Folder creation is possible, but we aren't going anywhere with that yet.
At one point the program was too complex for me to work on,
so I simplified it. There remains a lot of crust to remove.